MirageServer/requirements.txt
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
discord==2.3.1
discord.py==2.3.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
psutil==5.9.5
pyre-extensions==0.0.29
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
websockets==11.0.3
xformers==0.0.20
yarl==1.9.2
zipp==3.16.2
